requirements.txt
# FastAPI and ASGI server
fastapi
uvicorn[standard]
python-multipart

# Environment variables
python-dotenv

# Supabase client
supabase

# HTTP client for Clerk verification
httpx

# JWT verification
PyJWT[crypto]

# ML and Data Processing
numpy
pandas
scikit-learn
joblib
opencv-python
ultralytics
torch
Pillow
